% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cna.call}
\alias{cna.call}
\title{TCGA AML SNP6.0 GISTIC copy number alteration calls}
\format{A matrix containing the GISTIC copy number calls of 454 genes in 197 patients:
\itemize{
  \item{Row names are patient identifiers}
  \item{Colume names are official HGNC gene symbols}
}
Each element of the matrix is coded:
\itemize{
  \item -2, homozygous deletions
  \item -1, hemizygous deletions
  \item 0, neutral
  \item 1, gain
  \item 2, amplifications
  }}
\source{
\url{https://www.synapse.org/#!Synapse:syn300013}
}
\usage{
cna.call
}
\description{
A dataset containing part of the The Cancer Genome Atlas acute myeloid leukemia
Affymetrix SNP6.0 array copy number alteration calls from GISTIC
}
\keyword{datasets}

