% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cna.logr}
\alias{cna.logr}
\title{TCGA AML SNP6.0 copy number alteration data}
\format{A matrix containing the copy number log2 ratios of 454 genes in 197 patients:
\itemize{
  \item{Row names are patient identifiers}
  \item{Colume names are official HGNC gene symbols}
}}
\source{
\url{https://www.synapse.org/#!Synapse:syn300013}
}
\usage{
cna.logr
}
\description{
A dataset containing part of the The Cancer Genome Atlas acute myeloid leukemia
Affymetrix SNP6.0 array copy number alteration log2 ratios
}
\keyword{datasets}

