% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xseq_inference_learning.R
\name{ConvertXseqOutput}
\alias{ConvertXseqOutput}
\title{Convert xseq output to a data.frame}
\usage{
ConvertXseqOutput(posterior)
}
\arguments{
\item{posterior}{The posterior probabilities of mutations and mutated genes,
output from \code{InferXseqPosterior}}
}
\value{
A data.frame with sample, gene,
probability of individual mutations and the probabilities of individual mutated genes
}
\description{
Convert xseq output to a data.frame
}

