% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{expr}
\alias{expr}
\title{TCGA AML SNP6.0 gene expression data}
\format{A matrix containing the expression of 454 genes in 197 patients:
\itemize{
  \item{Row names are patient identifiers}
  \item{Colume names are official HGNC gene symbols}
}}
\source{
\url{https://www.synapse.org/#!Synapse:syn300013}
}
\usage{
expr
}
\description{
A dataset containing part of the The Cancer Genome Atlas acute myeloid leukemia RNA-seq gene expression data
}
\keyword{datasets}

