% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{ImputeKnn}
\alias{ImputeKnn}
\title{Impute missing values (NAs) using K-nearest neighbour averaging}
\usage{
ImputeKnn(X, ratio = 0.5, ...)
}
\arguments{
\item{X}{A matrix of real values where each row corresponds to a patient
and each column is a gene.}

\item{ratio}{The rows (columns) with more than (default 50\%) of
missing values are removed}

\item{...}{Arguments to be passed to impute.knn}
}
\value{
A matrix without NAs
}
\description{
Impute missing values (NAs) using K-nearest neighbour averaging
}
\examples{
data(expr)
expr.norm = ImputeKnn(expr)
}

