% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xseq_inference_learning.R
\name{InferXseqPosterior}
\alias{InferXseqPosterior}
\title{Learn xseq parameters given an initialized model}
\usage{
InferXseqPosterior(model, constraint, debug = FALSE)
}
\arguments{
\item{model}{An xseq model.}

\item{constraint}{A list of constraints on \eqn{\theta_{G|F}}.}

\item{debug}{Logical, specifying whether debug information should be printed}
}
\value{
The posterior probabilities of latent variables in xseq
}
\description{
Learn xseq parameters given an initialized model
}

