% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xseq_inference_learning.R
\name{LearnXseqParameter}
\alias{LearnXseqParameter}
\title{Learn xseq parameters given an initialized model}
\usage{
LearnXseqParameter(model, constraint, iter.max = 20, threshold = 1e-05,
  cis = FALSE, debug = FALSE, show.plot = TRUE)
}
\arguments{
\item{model}{An xseq model}

\item{constraint}{A list of constraints on \eqn{\theta_{G|F}}.}

\item{iter.max}{Maximum number of iterations in learning xseq parameters}

\item{threshold}{The threshold to stop learning paramters}

\item{cis}{Logical, cis-analysis or trans-analysis}

\item{debug}{Logical, specifying whether debug information should be printed}

\item{show.plot}{Logical, specifying whether to plot the Log-Likelihoods}
}
\value{
A list including the learned xseq model
}
\description{
Learn xseq parameters given an initialized model
}

