% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mut}
\alias{mut}
\title{TCGA AML somatic mutation data}
\format{A data frame with 2311 rows and 12 variables:
\itemize{
  \item sample. character, patient identifier
  \item hgnc_symbol. character, official HGNC gene symbols
  \item variant_type. character, mutation type, can be either "HOMD", "HLAMP",
  "MISSENSE", "NONSENSE", "FRAMESHIFT", "INFRAME", "SPLICE", "NONSTOP",
  "STARTGAINED", "SYNONYMOUS", "OTHER", "FUSION", "COMPLEX"
  \item ...
}}
\source{
\url{https://www.synapse.org/#!Synapse:syn1729383}
}
\usage{
mut
}
\description{
A dataset containing the The Cancer Genome Atlas acute myeloid leukemia somatic mutation data
}
\keyword{datasets}

