% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{net}
\alias{net}
\title{A networks containing gene associations}
\format{A list with 16 elements:
\itemize{
  \item{List names are official HGNC gene symbols}
  \item{Each element of the list is a vector, and the name of the vector is an official HGNC gene symbol.
  Each element of the vector is a real number between 0 and 1,
  representing the assocation strongth between two genes.
  The names of the elements of the vector are also official HGNC gene symbols.}
}}
\usage{
net
}
\description{
A list of gene interactions
}
\keyword{datasets}

