% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xseq_plot.R
\name{PlotRegulationHeatmap}
\alias{PlotRegulationHeatmap}
\title{Heatmap showing the connected genes' dysregulation probabilities}
\usage{
PlotRegulationHeatmap(gene, posterior, mut, subtype = NULL,
  main = "in_Cancer", ...)
}
\arguments{
\item{gene}{A character vector of gene names}

\item{posterior}{The xseq posteriors, output of \code{InferXseqPosterior}
or \code{LearnXseqParameter}}

\item{mut}{A data.frame of mutations. The data.frame should have at least
three columns of characters: sample, hgnc_symbol, and variant_type.
The variant_type column cat be either "HOMD", "HLAMP",
  "MISSENSE", "NONSENSE", "FRAMESHIFT", "INFRAME", "SPLICE", "NONSTOP",
  "STARTGAINED", "SYNONYMOUS", "OTHER", "FUSION", "COMPLEX".}

\item{subtype}{A vector representing a character of each patient,
e.g., subtype}

\item{main}{The heatmap title}

\item{...}{Other parameters passed to heatmap.2}
}
\description{
Heatmap showing the connected genes' dysregulation probabilities
}

