% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{QuantileNorm}
\alias{QuantileNorm}
\title{Quantile normalize a matrix}
\usage{
QuantileNorm(X)
}
\arguments{
\item{X}{A matrix of real values where each row corresponds to a patient
and each column is a gene}
}
\value{
The normalized matrix of \code{X}
}
\description{
Quantile normalize a matrix
}
\examples{
data(expr)
expr.quantile = QuantileNorm(expr)
}

