% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xseq_inference_learning.R
\name{SetXseqPrior}
\alias{SetXseqPrior}
\title{Set model paramerter priors}
\usage{
SetXseqPrior(mut, expr.dis, net, regulation.direction = TRUE, cis = TRUE,
  mut.type = "loss", ...)
}
\arguments{
\item{mut}{A data.frame of mutations. The data.frame should have
three columns of characters: sample, hgnc_symbol, and variant_type.
The variant_type column cat be either "HOMD", "HLAMP",
  "MISSENSE", "NONSENSE", "FRAMESHIFT", "INFRAME", "SPLICE", "NONSTOP",
  "STARTGAINED", "SYNONYMOUS", "OTHER", "FUSION", "COMPLEX".}

\item{expr.dis}{A list, the outputs from calling \code{GetExpressionDistribution}}

\item{net}{A list of gene interactions}

\item{regulation.direction}{Logical, whether considering the directionality
, i.e., up-regulation or down-regulatin of genes, only used
when \code{cis=FALSE}.}

\item{cis}{Logical, cis analysis or trans analysis}

\item{mut.type}{Character, only used when \code{cis = TRUE}, and can be either
loss, gain or both}

\item{...}{Reserved for extension}
}
\description{
Set model paramerter priors
}

